Verify user has access to visualization before creating widget
