+++ redash/handlers/widgets.py
@@ -4,7 +4,7 @@ from flask import request
 
 from redash import models
 from redash.wsgi import api
-from redash.permissions import require_permission, require_admin_or_owner
+from redash.permissions import require_permission, require_admin_or_owner, require_access, view_only
 from redash.handlers.base import BaseResource
 
 
@@ -18,8 +18,16 @@ class WidgetListAPI(BaseResource):
         widget_properties['options'] = json.dumps(widget_properties['options'])
         widget_properties.pop('id', None)
         widget_properties['dashboard'] = dashboard
-        # TODO: verify access to the visualization
-        widget_properties['visualization'] = widget_properties.pop('visualization_id')
+
+        visualization_id = widget_properties.pop('visualization_id')
+        if visualization_id:
+            visualization = models.Visualization.get_by_id_and_org(visualization_id, self.current_org)
+            require_access(visualization.query.groups, self.current_user, view_only)
+        else:
+            visualization = None
+
+        widget_properties['visualization'] = visualization
+
         widget = models.Widget.create(**widget_properties)
 
         layout = json.loads(widget.dashboard.layout)
